requirements.txt
# Python library requirements for the AutoAgendaWriter application.
# Use 'pip install -r requirements.txt' to install them.

# GUI Framework
Kivy==2.3.1
Pillow==11.3.0  # For image support in Kivy

# Data Processing
pandas==2.3.1
openpyxl==3.1.5 # For reading .xlsx files
numpy==2.3.1    # Core dependency for pandas

# Clipboard fallback (optional but recommended for Windows)
pyperclip

# LLM and AI
# LLM and AI
# IMPORTANT FOR WINDOWS USERS:
# Installing llama-cpp-python can be complex. For GPU acceleration (NVIDIA),
# it is highly recommended to install it manually *before* running `pip install -r requirements.txt`.
# Open a command prompt and run these commands, replacing with your CUDA version if needed:
#   set CMAKE_ARGS="-DLLAMA_CUBLAS=on"
#   set FORCE_CMAKE=1
#   pip install llama-cpp-python --no-cache-dir --force-reinstall
# If you do not have an NVIDIA GPU or do not want GPU support, the standard pip install may work,
# but you might need the Visual Studio Build Tools installed.
llama-cpp-python==0.3.13
huggingface-hub==0.33.4  # For downloading models from Hugging Face Hub

# Document Generation
python-docx==1.2.0

# For system notifications and other platform-specific features
plyer
win10toast; sys_platform == 'win32'  # For Windows notifications via plyer